Performance issue
Seeing the game in a webgl build and it seems the performance is horrible.
In terms of FPS, it's fine, but it has a 4ms delay, which is actually quite bad.
Source of this lag is clearly coming from the lines, I'm not really sure how to optimize it, but I'll figure something out.